bank account class

+++ src/ValidNumbers.java
@@ -2,15 +2,15 @@ import java.util.Scanner;
 
 public class ValidNumbers {
 
-    public static int validateAndSumDigits() {
-        int sum = 0;
+    public static double validateAndSumDigits() {
+        double sum = 0;
         int count = 1;
         Scanner input = new Scanner(System.in);
 
         do {
             try {
                 System.out.println("Enter number #" + count + ": ");
-                sum += input.nextInt();
+                sum += input.nextDouble();
                 count++;
             } catch (Exception e) {
                 System.out.println("Invalid input. Please enter a valid number.");
